[api/debug] add debug home
