+++ fusequery/query/src/api/http/debug/mod.rs
@@ -2,4 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0.
 
+pub mod home;
 pub mod pprof;
+
+pub use home::PProfRequest;
