+++ fusequery/query/src/api/http/router.rs
@@ -23,8 +23,8 @@ impl Router {
     ) -> Result<impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone> {
         let v1 = super::v1::hello::hello_handler(self.cfg.clone())
             .or(super::v1::config::config_handler(self.cfg.clone()))
-            .or(super::debug::pprof::pprof_handler(self.cfg.clone()))
-            .or(super::v1::cluster::cluster_handler(self.cluster.clone()));
+            .or(super::v1::cluster::cluster_handler(self.cluster.clone()))
+            .or(super::debug::home::debug_handler(self.cfg.clone()));
         let routes = v1.with(warp::log("v1"));
         Ok(routes)
     }
